refactor(harness): extract command registry framework with configurable skill loader
