+++ packages/cea/src/commands/aliases-and-tool-fallback.test.ts
@@ -1,4 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it } from "bun:test";
+import { type Command, createHelpCommand } from "@ai-sdk-tool/harness";
 import { agentManager } from "../agent";
 import {
   DEFAULT_TOOL_FALLBACK_MODE,
@@ -6,11 +7,9 @@ import {
   TOOL_FALLBACK_MODES,
   type ToolFallbackMode,
 } from "../tool-fallback-mode";
-import { createHelpCommand } from "./help";
 import { executeCommand, getCommands, registerCommand } from "./index";
 import { createReasoningModeCommand } from "./reasoning-mode";
 import { createToolFallbackCommand } from "./tool-fallback";
-import type { Command } from "./types";
 
 let commandCounter = 0;
