+++ packages/cea/src/commands/reasoning-mode.ts
@@ -1,6 +1,6 @@
+import type { Command } from "@ai-sdk-tool/harness";
 import { agentManager } from "../agent";
 import { parseReasoningMode, REASONING_MODES } from "../reasoning-mode";
-import type { Command } from "./types";
 
 const REASONING_MODE_USAGE = REASONING_MODES.join("|");
